Windows build.

+++ src/tree/gpu_hist/histogram.cu
@@ -203,7 +203,8 @@ void BuildGradientHistogram(EllpackDeviceAccessor const& matrix,
   grid_size = common::DivRoundUp(grid_size,
       common::DivRoundUp(num_groups, num_groups_threshold));
 
-  dh::LaunchKernel {dim3(grid_size, num_groups), block_threads, smem_size} (
+  dh::LaunchKernel {
+    dim3(grid_size, num_groups), static_cast<uint32_t>(block_threads), smem_size} (
       kernel,
       matrix, feature_groups, d_ridx, histogram.data(), gpair.data(), rounding,
       shared);
